refactor: 이슈 패키지 리팩토링
- DTO 명칭 일관성있게 변경
- mapping 메서드 mapper클래스에서 수행하도록 변경
