+++ be/issue-tracker/src/main/java/com/issuetracker/issue/dto/IssueCreateRequest.java
@@ -9,7 +9,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 @Getter
 @Builder
-public class IssueCreateRequestDto {
+public class IssueCreateRequest {
     @NotBlank
     private final String title;
     @NotBlank
